Ensure live map only processes Facepunch level URLs
